Add files via upload
1) Wind Vane

This circuit can detect wind speed and wind load (we could measure pressure instead if we knew the exact area of the sensor). For measuring the direction, I wanted a sensor that can rotate and change some parameter to reflect the angle of rotation, so I used the circular potentiometer. A tall wind vane will be mounted on this, and when that rotates the potentiometer knob will rotate as well. I am using its output voltage to find the angle covered using simple ratios, as the voltage varies linearly based on the arc length that the potentiometer knob has rotated by. To measure the load, I used simple ratio (voltage from sensor at the instant/max voltage==force applied at the instant/maximum force that it can measure). The direction and load are displayed on the LCD for the user to see.

Limitations/Improvements:
- The circular potentiometer on Tinkercad can only rotate 270 degrees and not 360, so I'm measuring angles only from 45 to 135. In real life, I would use a potentiometer that can rotate as close to 360 degrees as possible.
- Since I could not find the dimensions of the FSR on Tinkercad, I could not find pressure. In real life, we can find the pressure using a given area and find the wind velocity from that and average density of air.
- The force sensor has a fixed sensitivity. To change that, I would use a potentiometer that would let the user decide the sensitivity based on the location or application.
----------------------------------------------------------------------------------------
2) Plotting the effect of temperature on soil moisture level

This circuit measures the temperature and soil moisture level, graphs the instantaneous change in soil moisture level over time using an oscilloscope (its slope gives evaporation rate) and displays the temperature and instantaneous evaporation rate at that time on the LCD. To find the rate of evaporation, I found the change in moisture level between two runs of void loop() (which has a delay of 1s) using a variable that keeps track of time and adds 1 second every run. Then I subtracted the previous second's moisture level from the current second's moisture level to get the rate of change of water i.e. rate of evaporation. I thought the oscilloscope would be a nice way of displaying how moisture changes with time at any given temperature, so I used that.
----------------------------------------------------------------------------------------------
3) Smoke management system

While the slide switch is on, the system is operational.
If the smoke level is below a first threshold level, nothing happens.
If the smoke level is above the first threshold level but below a higher second threshold level, the exhaust fan works.
If the smoke level is above the second threshold level, an alarm is sounded using the buzzer.